Move files into subfolders and dedicated namespaces
