--- inc/ServiceProvider.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace DebugHooks;
-
-class ServiceProvider extends Dependencies\LaunchpadAutoresolver\ServiceProvider
-{
-    public function get_common_subscribers(): array {
-        return [
-            MySubscriber::class,
-        ];
-    }
-}
